Moving header and footer back out
