+++ src/Components/Footer/Footer.js
@@ -1,4 +1,4 @@
-import "../../../App.css";
+import "../../App.css";
 
 const Footer = () => <h2>Thank you for visiting!</h2>;
