+++ src/Components/Header/Header.js
@@ -1,4 +1,4 @@
-import "../../../App.css";
+import "../../App.css";
 
 const Header = () => <h1>Matthew Nobes</h1>;
